Inference v2
This model was trained on uniform size 100x100, whereas previously, we had a range of cropped sizes. Why did I do this ? 1. Because I can. 2. But really, it was because the new setup we have on the 64-bit Pi OS with the libcamera module (we previously used the PiCamera on 32-bit) gives a range of cropped images that is larger than the one the original model was trained on, so it basically was giving us random classifications (1/6 in chance). Hopefully this fixes things.